Add sample multi-release setup in spring-context
NOTE: this commit is not meant to be pushed as-is in the main branch,
this is merely a sample showing how the multi-release Jar support can be
configured for a Spring Framework module.

This commit configures the multi-release Jar support in spring-context,
with a JDK17 baseline and a JDK 21 variant.
New `VirtualThreadsTaskExecutor` implementations are contributed:

* a JDK17 baseline that does not support Virtual Threads and throws an
  exception.
* a JDK21 variant that "supports" Virtual Threads (here, doing nothing).

Both variants have dedicated tests; we must ensure that baseline tests
are ignored for JDK21+ because the multi-release variant will be picked
up if we test the project with "-PtestToolchain=21".