Add JMS MessageInterceptor
This commit adds a new contract `MessageInterceptor` that allows to
intercept messages when received or when they're about to be sent.
Implementations can mutate the `Message` or signal whether the message
should be ignored and further processing cancelled.

Interceptors can be configured on `JmsTemplate` for intercepting send
and receive operations. `AbstractMessageListenerContainer`
implementations can get configured with interceptors for receive
operations only.