Add execution information for Task and ScheduledTask
TODO: this implementation is causing failures in
`ScheduledTasksBeanDefinitionParserTests` and
`ScheduledAnnotationBeanPostProcessorTests` because they rely on the
fact that `Task.getRunnable()` should return the original type that was
initially set in the task.
Should we honor this behavior, making any wrapping/delegation impossible
here to track the execution state of the task?
Should we wrap the given `Runnable` and only honor the documented
contract?

Prior to this commit, both  `Task` and `ScheduledTask` did not expose
any information about the state of their execution.

With this commit, `ScheduledTask` can now return its `nextExecution` as
an `Instant` if it's currently scheduled.
This also make `Task` expose its `TaskExecutionOutcome`, an immutable
object tracking whether the task has been executed, when and the
exception thrown from the execution (if any).

Fixes gh-